Move optparser in front of curses. Add windowing. Make persistence more realistic.
Move the option parser outside the curses wrapper, so it works with the -h help
messages.

Change persistence so that is stacks PSF, rather than stacking antenna locations
(in real-life, this would be adding points on a uv plane corresponding to baselines
between an antenna and itself (and others) at previous times.

Reduce the gaussian width for prettiness, since the above change makes the PSF
worse.

Add experimental windowing function. Currently there's a bug somewhere that means
when windowing is on the scale of the PSF changes. Perhaps fix this, and change
the entire code to stack baseline distribution rather than PSF?

Misc changes to Hough circle detection. Should probably just make these a command line
parameter.